updated DateFormatting test to Swift
